small refactor for copy blocks ffi

+++ csrc/src/ffi.rs
@@ -66,7 +66,7 @@ extern "C" {
         block_mapping: *const c_void,
         num_layers: i64,
         num_pairs: i64,
-        numel_per_block: u32,
+        numel_per_block: i64,
         stream: *mut c_void,
     );
 
@@ -76,7 +76,7 @@ extern "C" {
         block_mapping: *const c_void,
         num_layers: i64,
         num_pairs: i64,
-        numel_per_block: u32,
+        numel_per_block: i64,
         stream: *mut c_void,
     );
